Remove the dependency on Git bash
Different Git packages use different installation paths, which, in
their turn may also be redefined by the user. Moreover, even if Git
is installed there is no guarantee that the Bash component is also
present. So there is little sense to depend on Git bash or to consider
it to be at any specific location.

Resolves: #343, #335, #640, #466, #687, #636.

Signed-off-by: Maxim Plevako <mplevako@gmail.com>