Change the way a Git installation is detected
Different Git packages use different installation paths, which, in
their turn, may also be redefined by the user. Also, many packages
don't modify the registry and even those that do use different registry
branches, entries and names. Therefore it is unreliable to check for
existence of any assumed paths, files or registry entries to detect
whether Git is already installed.
Docker itself assumes that the path to the git command is present in
the PATH environment variable so it is enough to use the where command
which is present in all Windows OSes required by Docker.

Resolves: #226, #246, #295, #276.

Signed-off-by: Maxim Plevako <mplevako@gmail.com>